% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.avg.exg.rate}
\alias{get.avg.exg.rate}
\title{Average Exchange Rate - THB / Foreign Currency}
\usage{
get.avg.exg.rate(id, start_period, end_period, freq = "d")
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{freq}{(optional) Frequency of data including four options: \cr
daily ('d'), monthly ('m'), quarterly ('q'), and annual ('y') \cr
Default as daily}
}
\value{
List of average exchange rate and their information
}
\description{
Get average exchange rate - THB / Foreign Currency \cr 
BoT reference: The data collected includes the exchange rates quoted for immediate delivery (i.e., in the spot market) between Thai Baht vis攼㸰-vis 48 other currencies. Exchange rate data are obtained from daily foreign exchange transaction reports from some commercial banks while some exchange rate data are collected from New York Closing and Financial Times. All rates would be converted to Thai Baht equivalent using cross rates quoted by Bangkok Market Closing. \cr
Data is calculated from Daily Intebank Purchases and Sales of US Dollar (against THB) of the transaction worth more than or equal to 1 million USD. The Exchange Rates are calculated using weighted-average between the trading volume and the exchange rate specified.
Release schedule : Every business day at 6.00 p.m. (BKK-GMT+07:00) \cr
Source of data : 1. Commercial Banks registered in Thailand 2. Foreign Bank Branches 3. Special-purpose Financial Institutions
}
\examples{

m.avg.ref.rate <- get.avg.exg.rate(id,'2017-01-01','2017-12-31', freq = 'm');

}
