% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.bibor.rates}
\alias{get.bibor.rates}
\title{Bangkok Interbank Offered Rate (BIBOR) (Percent per annum)}
\usage{
get.bibor.rates(id, start_period, end_period, bank = NULL, type = "raw")

get.bibor.rates(id, start_period, end_period, bank = NULL, type = "raw")
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{term_type}{Term type}
}
\value{
List of series of swap point - onshore and their information

List of series of swap point - onshore and their information
}
\description{
Get series of Bangkok international bank offered rate  \cr 
BoT reference: Interbank transactions rates are classified by lending period (tenor) such as Overnight (O/N) , Tom/Next , Overnight Forward Start , At Call and Fixed Term. \cr
Release schedule : every business day at 10.00 a.m.\cr
Lag time : 1 business day \cr
Source of data: All commercial banks, Export-Import Bank of Thailand , The Government Savings Bank , Bank for Agricultue and Agricultural Cooperatives , Small and Medium Enterprise Development Bank of Thailand , The Government Housing Bank and Islamic Bank of Thailand, and All finance companies

Get series of Bangkok international bank offered rate  \cr 
BoT reference: Interbank transactions rates are classified by lending period (tenor) such as Overnight (O/N) , Tom/Next , Overnight Forward Start , At Call and Fixed Term. \cr
Release schedule : every business day at 10.00 a.m.\cr
Lag time : 1 business day \cr
Source of data: All commercial banks, Export-Import Bank of Thailand , The Government Savings Bank , Bank for Agricultue and Agricultural Cooperatives , Small and Medium Enterprise Development Bank of Thailand , The Government Housing Bank and Islamic Bank of Thailand, and All finance companies
}
\examples{

bibor.rates  <- get.bibor.rates(id,'2017-01-01','2017-12-31');


bibor.rates  <- get.bibor.rates(id,'2017-01-01','2017-12-31');

}
