% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.category.list}
\alias{get.category.list}
\title{Stat Category}
\usage{
get.category.list(id)
}
\arguments{
\item{id}{Client ID obtained from BoT}
}
\value{
List of all data categories
}
\description{
Get all data categories \cr 
BoT reference: Series List Users can call operation to browse all available categories in the system. The system will display the category, description th and description eng in order for users to apply the desired category to search the Series List.
}
\examples{

list.cat <- get.category.list(id)

}
