% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.debt.sec}
\alias{get.debt.sec}
\title{Debt Securities Auction Result (Current)}
\usage{
get.debt.sec(id, start_period, end_period)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}
}
\value{
List of series of loan interest rates of commercial banks and their information
}
\description{
Get series of debt securities auction result  \cr 
BoT reference: Provide the result of Debt securities auction.\cr
Release schedule : when the auction is approved \cr
Frequency : Following to the auction schedule \cr
Source of data: Bank of Thailand \cr
Note: BOT has provided more information, such as Classification of Financial Instrument code (cfi_code) and Auction Status (auction_st), on API responses starting September 29, 2017 onward.
}
\examples{

debt.sec <- get.debt.sec(id,'2018-01-01','2018-12-31');
}
