% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.deposit.rates}
\alias{get.deposit.rates}
\title{Deposit Interest Rates for Individuals of Commercial Banks (Percent per annum)}
\usage{
get.deposit.rates(id, start_period, end_period, type = "raw")
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{type}{Type of obtained data: raw data ('raw') and average data ('avg); Default as 'raw'}
}
\value{
List of series of deposit interest rates for individuals of commercial banks and their information
}
\description{
Get series of deposit interest rates for individuals of commercial banks  \cr 
BoT reference: Refers to the announced deposit interest rate for individual of commercial banks, classified by deposit type such as saving and fixed term deposit. \cr
Release schedule : Every business day at 2.00 p.m.\cr
Frequency : Daily \cr
Source of data: Commercial Banks registered in Thailand and Foreign Bank Branches
}
\examples{

deposit.rates  <- get.deposit.rates(id,'2017-01-01','2017-12-31');
deposit.rates.avg  <- get.deposit.rates(id,'2017-01-01','2017-12-31','avg');
}
