% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.exg.rate}
\alias{get.exg.rate}
\title{Weighted-average Interbank Exchange Rate - THB / USD}
\usage{
get.exg.rate(id, start_period, end_period, freq = "d")
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{freq}{(optional) Frequency of data including four options: \cr
daily ('d'), monthly ('m'), quarterly ('q'), and annual ('y') \cr
Default as daily}
}
\value{
List of weighted-average international bank exchange rate and their information
}
\description{
Get weighted-average international bank exchange rate 
BoT reference: Daily Weighted-average Interbank Exchange Rate - THB / USD \cr
Data is calculated from Daily Intebank Purchases and Sales of US Dollar (against THB) of the transaction worth more than or equal to 1 million USD. The Exchange Rates are calculated using weighted-average between the trading volume and the exchange rate specified.
Release schedule : Every business day at 6.00 p.m. (BKK-GMT+07:00) \cr
Source of data : 1. Commercial Banks registered in Thailand 2. Foreign Bank Branches 3. Special-purpose Financial Institutions \cr
}
\examples{

m.get.exg.rate <- get.exg.rate(id,'2017-01-01','2017-12-31', freq = 'm');

}
