% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.ext.rates}
\alias{get.ext.rates}
\title{External Interest Rates (Percent per annum)}
\usage{
get.ext.rates(id, start_period, end_period, rate_type = NULL)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{rate_type}{(optional) Rate Type EX. Libor 1 W in USD (Avg.offered rates)}
}
\value{
List of external interest rates and their information
}
\description{
Get series of external interest rates\cr 
BoT reference: External Interest Rates encompass major interest rates \cr
}
\examples{

ext.rates <- get.ext.rates(id,'2017-01-01','2017-12-31');

}
