% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.holiday}
\alias{get.holiday}
\title{Financial Institutions' Holidays}
\usage{
get.holiday(id, year)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{year}{Year (Ex. 2021)}
}
\description{
Get series of financial institutions' holidays in given year
}
\examples{

holiday <- get.holiday(id,'2021');
}
