% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.inter.tran.rates}
\alias{get.inter.tran.rates}
\title{Interbank Transaction Rates}
\usage{
get.inter.tran.rates(id, start_period, end_period, term_type = NULL)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{term_type}{Term type}
}
\value{
List of series of swap point - onshore and their information
}
\description{
Get series of International bank transaction rates \cr 
BoT reference: Swap Point Onshore (in Satangs) refers to the difference between the spot and forward exchange rates available at the period of 1, 3 and 6 months. \cr
Release schedule : Interbank transactions rates are classified by lending period (tenor) such as Overnight (O/N) , Tom/Next , Overnight Forward Start , At Call and Fixed Term. \cr
Lag time : 1 business day \cr
Release schedule : every business day at 10.00 a.m. \cr
Source of data: All commercial banks, Export-Import Bank of Thailand , The Government Savings Bank , Bank for Agricultue and Agricultural Cooperatives , Small and Medium Enterprise Development Bank of Thailand , The Government Housing Bank and Islamic Bank of Thailand, and All finance companies
}
\examples{

inter.tran.rates  <- get.inter.tran.rates(id,'2017-01-01','2017-12-31');

}
