% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.loan.rates}
\alias{get.loan.rates}
\title{Loan Interest Rates of Commercial Banks (Percent per annum)}
\usage{
get.loan.rates(id, start_period, end_period, type = "raw")
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{type}{Type of obtained data: raw data ('raw') and average data ('avg); Default as 'raw'}
}
\value{
List of series of loan interest rates of commercial banks and their information
}
\description{
Get series of loan interest rates of commercial banks (percent per annum)  \cr 
BoT reference: Refers to the announced loan interest rate of commercial banks, classified by loan type such as MOR (Minimun Overdraft Rate), MLR (Minimun Loan Rate) and MRR (Minimun Retail Rate) etc.\cr
Release schedule : Every business day at 2.00 p.m.\cr
Frequency : Daily \cr
Source of data: Commercial Banks registered in Thailand and Foreign Bank Branches
}
\examples{

loan.rates <- get.loan.rates(id,'2017-01-01','2017-12-31');
loan.rates.avg <- get.loan.rates(id,'2017-01-01','2017-12-31','avg');
}
