% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.observe}
\alias{get.observe}
\title{Observations}
\usage{
get.observe(id, series_code, start_period, end_period = NULL, sort_by = NULL)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{series_code}{Code of data series provided by BoT}

\item{start_period}{Star period: format (YYYY-MM-DD)}

\item{end_period}{(optional) End period: format (YYYY-MM-DD)}

\item{sort_by}{(optional) Sort data ('asc' or 'desc')}
}
\value{
List of series of data
}
\description{
Batching Data by given series code and initial date \cr 
BoT reference: The BOT API products on Statistics include economic and financial, financial Institutions, financial market and payment systems statistics, all of which are published on the BOT Website. Users can filter observations by series code, downloadable from List of Statistics APIs below. \cr 
https://www.bot.or.th/Thai/Statistics/API_DOC/API%20Statistic%20time%20series.xlsx
}
\examples{

series_code <- 'PF00000000Q00232'
start_period <- '2017-10-01'
end_period <- '2017-11-30'
sort_by <- 'desc'

FSI <- get.observe(id, series_code, start_period);

FSI.end <- get.observe(id, series_code, start_period, end_period, sort_by);

}
