% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.policy.rate}
\alias{get.policy.rate}
\title{Policy Rate (Percent per annum)}
\usage{
get.policy.rate(id)
}
\arguments{
\item{id}{Client ID obtained from BoT}
}
\value{
Current policy rate
}
\description{
Get current policy rate \cr 
BoT reference: Policy Interest Rate is the rate that The Monetary Policy Committee announced in conducting monetary policy under the inflation-targeting framework. The 14-day RP rate was used as the policy interest rate up until 16 January 2007, after which the policy interest rate was switched to the 1- day RP rate. Since 12 February 2008, with the closure of the BOT- run RP market, this was switched to the 1-day bilateral RP rate. \cr
Release schedule : Interbank transactions rates are classified by lending period (tenor) such as Overnight (O/N) , Tom/Next , Overnight Forward Start , At Call and Fixed Term. \cr
Frequency : Following to the schedule of The Monetary Policy Committee Meeting  \cr
Release schedule :  2.05 p.m. of the meeting date \cr
Source of data: Bank of Thailand
}
\examples{

inter.tran.rates  <- get.inter.tran.rates(id,'2017-01-01','2017-12-31');

}
