% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.product}
\alias{get.product}
\title{Financial Product}
\usage{
get.product(id, data = NULL, type = "deposit")
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{data}{(optional) Specific data of the product, see detail at Bot API page}

\item{type}{Type of financial product including six options: \cr 
Deposit ('deposit'), Personal Loan ('ploan'), Credit Card('credit'), Debit Card ('debit'), and Home Loan ('hloan'); Default setting as 'deposit'}
}
\value{
List of productions
}
\description{
Download list of financial product template that financial institutions are required to submit updated information to BOT one day before effective date
}
\examples{

deposit <- get.product(id);

data <- list("FICodeList" = "002", \cr 
"AccountTypeList" = "1",\cr 
 "BalanceAmount" = "",\cr 
  "ProductName" = "", \cr 
  "DepositTermRange" = "",\cr 
   "DepositTermType" = "", \cr 
   "InterestWithoutTax" = "",\cr 
   "AccntWithInsrnc" = "",\cr 
    "ProductRelate" = ""); 
deposit.data <- get.product(id, data);

ploan <- get.product(id, type = 'ploan');

}
