% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.series.list}
\alias{get.series.list}
\title{Stat Series}
\usage{
get.series.list(id, category)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{category}{category for get all series in category}
}
\value{
List of all data series in given category
}
\description{
Get all data series in given category \cr 
BoT reference: Users can define the category they want by entering the code of the category, for example EC_XT_077. The system will display a list of all series in the Category so that users can apply the series list to retrieve the required information in APIs (Observations).
}
\examples{

list.series <- get.series.list(id,'EC_EI_001_S2');

}
