% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.spot.rates}
\alias{get.spot.rates}
\title{Spot Rate USD/THB}
\usage{
get.spot.rates(id, start_period, end_period)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}
}
\value{
List of spot rate in USD/THB and their information
}
\description{
Get series of spot rate in USD/THB \cr 
BoT reference: Spot Rate : USD/THB refers to the exchange rate quoted on the spot market. All trading transactions are verified and settled within two consecutive business day following the trading date. \cr
Release schedule : End of every business day at 6.00 p.m.
}
\examples{

spot.rates <- get.spot.rates(id,'2017-01-01','2017-12-31');

}
