% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.swap.rates}
\alias{get.swap.rates}
\title{Swap point - Onshore (in Satangs)}
\usage{
get.swap.rates(id, start_period, end_period, term_type = NULL)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{term_type}{Term type}
}
\value{
List of series of swap point - onshore and their information
}
\description{
Get series of swap point - onshore in Satangs (Smallest Thai currency unit) \cr 
BoT reference: Swap Point Onshore (in Satangs) refers to the difference between the spot and forward exchange rates available at the period of 1, 3 and 6 months. \cr
Release schedule : End of every business day at 6.00 p.m. \cr
Source of data: Bank of Thailand and Reuters
}
\examples{

swap.rates  <- get.swap.rates(id,'2017-01-01','2017-12-31');

}
