% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{get.tbimplied.rates}
\alias{get.tbimplied.rates}
\title{External Interest Rates (Percent per annum)}
\usage{
get.tbimplied.rates(id, start_period, end_period, rate_type = NULL)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{start_period}{Start Period Date (YYYY-MM-DD)}

\item{end_period}{End Period Date (YYYY-MM-DD)}

\item{rate_type}{(optional) Rate Type EX. Libor 1 W in USD (Avg.offered rates)}
}
\value{
List of Thai Baht implied interest rates and their information
}
\description{
Get series of Thai Baht implied interest rates \cr 
BoT reference: Thai baht Implied Interest Rates refer to interest rate quoted for Baht borrowing through swap market from which the Bank of Thailand disseminates. The rates are those pertaining to domestic market where all transactions are either interbank or between commercial banks and their clients, and interest rates pertaining to foreign markets, which reflect transactions between commercial banks and Non-resident clients. \cr
 Frequency : Daily \cr
 Lag time : 1 \cr
 business; day \cr 
 Release schedule : Every business day at 11.00 a.m. \cr
 Source of data : Bank of Thailand
}
\examples{

tbimplied.rates <- get.tbimplied.rates(id,'2017-01-01','2017-12-31');

}
