% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOTapi.R
\name{search.stat}
\alias{search.stat}
\title{Search Stat APIs}
\usage{
search.stat(id, keyword)
}
\arguments{
\item{id}{Client ID obtained from BoT}

\item{keyword}{keyword for search series}
}
\value{
List of series information
}
\description{
Search data series information by a keyword \cr 
BoT reference: User can look for data in BOT API Statistics by keyword search using series code, series name, or relevant terms. The result will be displayed up to 100 series per search.
}
\examples{

income.series <- search.stat(id,'income') 

}
